server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psutil==6.0.0
pytest==8.3.2
paho-mqtt==2.1.0
httpx==0.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
